Create the /api/users endpoint
In this commit, I created the /api/users endpoint which returns a list of all users present in the database.
In achieving this, we make a database call, using the helper function within db.js called find(). The resolved promise, which is an array of objects containing each user is then returned