Create an endpoint for creating new user
In this commit, i created an endpoint /api/users that handles
post requests for creating new user. when a request hits this
post route, the db.create() function, which takes the new user as arg, is then invoked and then a new user is created. The id of that new user is returned in the .then() block