Create a new endpoint for updating a user
In this commit, I created a new put route /api/users/:id for updating user information.
This route when triggered, takes the id of the user and the user info with which it would update the users' details.
within the updateUserById callback function, the id and the user object is then used to make a put() request to the database for updating that particular user's information.