Create a delete endpoint
In this commit, I created a delete endpoint for deleting a particular user.
This API endpoint receives delete request from the client side.
when a request matches the delete('/api/users/:id') route, a callback handler function called deleteUserById() is triggered.
this function uses the id that was passed into it to call a remove() helper method available in db.js to query the database and delete any user which matches that particular id